+++ Content.Server/Remotes/DoorRemoteSystem.cs
@@ -74,7 +74,7 @@ namespace Content.Server.Remotes
             }
 
             if (TryComp<AccessReaderComponent>(args.Target, out var accessComponent) &&
-                !_doorSystem.HasAccess(args.Target.Value, args.Used, accessComponent))
+                !_doorSystem.HasAccess(args.Target.Value, args.Used, doorComp, accessComponent))
             {
                 _doorSystem.Deny(args.Target.Value, doorComp, args.User);
                 ShowPopupToUser("door-remote-denied", args.User);
